app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiogram==3.4.1
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
qrcode==7.4.2
Pillow==10.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psycopg-binary==3.1.19
